test: move biometric tests to patrol_test/biometric with local tag
Split the success and cancel scenarios into separate files under
patrol_test/biometric/ and tag them 'local' instead of
'android'/'emulator'. CI selects tests with --tags='android && emulator'
and runs them on emulator.wtf, a cloud device farm where biometric
enrollment cannot work (it needs patrol_cli driving the emulator console
from the host). The 'local' tag keeps these tests out of that selection —
they are meant for local emulator runs only.
